requirements.txt
<<<<<<< HEAD
streamlit==1.45.1
pandas==2.2.3
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.5.2
openai==1.93.2
pyyaml==6.0.2
bcrypt==4.3.0
requests==2.32.4
cryptography==45.0.5
pyjwt==2.10.1
psycopg[binary]==3.2.9
openpyxl==3.1.5
plotly==5.24.1
tensorflow==2.16.2
python-dotenv==1.1.1
=======
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.35.42
botocore==1.35.99
cachetools==5.5.2
captcha==0.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
extra-streamlit-components==0.1.80
Flask==3.1.1
fonttools==4.58.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.41.0
numpy==1.26.4
openai==1.82.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotly==5.24.1
protobuf==4.25.3
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
statsmodels==0.14.4
streamlit==1.45.1
streamlit-authenticator==0.4.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
>>>>>>> 41c87d72dd7eb52f51936fb2df2f7a4a7b87eb66
